requirements.txt
discord.py==2.3.2
APScheduler==3.10.4
